requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
PyYAML==6.0.2
python-dotenv==1.1.1
httpx==0.28.1
python-multipart==0.0.20
pydantic[email]==2.11.7
pytest==8.4.1
